moving function animateCSS to helper.js
