+++ src/script.js
@@ -1,4 +1,5 @@
 import { WORDS } from './words.js';
+import { animateCSS } from './helper.js';
 
 const openHintBtn = document.querySelector('.open-hint-section-btn');
 const inputQuery = document.getElementById('hint-input');
@@ -11,26 +12,6 @@ let guessesRemaining = NUMBER_OF_GUESSES;
 let currentGuess = [];
 let nextLetter = 0;
 let rightGuessString = WORDS[Math.floor(Math.random() * WORDS.length)];
-
-const animateCSS = (element, animation, prefix = 'animate__') =>
-  // We create a Promise and return it
-  new Promise((resolve, reject) => {
-    const animationName = `${prefix}${animation}`;
-    // const node = document.querySelector(element);
-    const node = element;
-    node.style.setProperty('--animate-duration', '0.3s');
-
-    node.classList.add(`${prefix}animated`, animationName);
-
-    // When the animation ends, we clean the classes and resolve the Promise
-    function handleAnimationEnd(event) {
-      event.stopPropagation();
-      node.classList.remove(`${prefix}animated`, animationName);
-      resolve('Animation ended');
-    }
-
-    node.addEventListener('animationend', handleAnimationEnd, { once: true });
-  });
 
 function initBoard() {
   const board = document.getElementById('game-board');
